+++ packages/apollo-angular/src/Apollo.ts
@@ -12,7 +12,7 @@ import {Observable} from 'rxjs/Observable';
 import {from} from 'rxjs/observable/from';
 
 import {QueryRef} from './QueryRef';
-import {TypedVariables, R} from './types';
+import {TypedVariables, ExtraSubscriptionOptions, R} from './types';
 import {fromPromise, wrapWithZone} from './utils';
 
 export class ApolloBase<TCacheShape = any> {
@@ -40,8 +40,13 @@ export class ApolloBase<TCacheShape = any> {
     );
   }
 
-  public subscribe(options: SubscriptionOptions): Observable<any> {
-    return wrapWithZone(from(this.client.subscribe({...options})));
+  public subscribe(
+    options: SubscriptionOptions,
+    extra?: ExtraSubscriptionOptions,
+  ): Observable<any> {
+    const obs = from(this.client.subscribe({...options}));
+
+    return extra && extra.useZone !== true ? obs : wrapWithZone(obs);
   }
 
   public getClient() {