+++ packages/apollo-angular/src/types.ts
@@ -3,3 +3,7 @@ export type R = Record<string, any>;
 export type TypedVariables<T> = {
   variables?: T;
 };
+
+export interface ExtraSubscriptionOptions {
+  useZone?: boolean;
+}