+++ packages/apollo-angular/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### vNEXT
 
+* Introduces `ExtraSubscriptionOptions`. Allow to run `Apollo.subscribe` outside
+  the Zone.
+* Exposes `wrapWithZone`, in case someone wants to run `Observable` inside the
+  Zone.
+
 ### v1.0.1
 
 * Brings typed variables to `QueryRef`